papers/QRNN/requirements.txt
torch>=2.2
pandas>=2.1
kagglehub>=0.2.5
numpy>=1.26
matplotlib>=3.7
merlinquantum
pytest

